requirements.txt
kafka-python
pandas
pyarrow
pymongo
cassandra-driver
fastparquet
confluent-kafka